Add BucketSize parameter.
For small N computing the prefix sum directly can be more efficient than via the tree (due to vectorization and the overhead from traversing the tree).
To optimize performance for such N, we can increase the threshold for the minimum number of elements a node can represent. Originally, it was 3, but
with the new parameter it becomes 2 * BucketSize (except the last node, which may represent just BucketSize+1 elements).
BucketSize=2 is optimal if addition is much more expensive than traversing the tree.
However, for simple arithmetic types larger BucketSize give better performance (e.g., 128 gives better performance for uint32 on my machine).