Use CompressedPath to construct FullTreeView.
Constructing a FullTreeView for the currently effective tree requires
determining the level (depth) of that tree and its bucket capacity.
While doing both takes O(1) time, it requires multiple arithmetic instructions,
thus affecting the constant in the time complexities of increment(),
prefixSum(), etc. This constant can be reduced if we reuse the data in
CompressedPath to build the tree view.